Clean up helper functions
